LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/document-exit-fullscreen-nested-shadow-dom-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
fullscreen

Harness Error (FAIL), message = Unhandled rejection: Type error

TIMEOUT Exiting fullscreen from a nested shadow root works correctly. Test timed out

Harness Error (FAIL), message = Unhandled rejection: Type error

TIMEOUT Exiting fullscreen from a nested shadow root works correctly. Test timed out
